snowflake - manifest load support
